requirements.txt
absl-py==0.9.0
alabaster==0.7.12
amqp==1.4.9
anyjson==0.3.3
appdirs==1.4.3
asgiref==3.2.7
astor==0.8.1
astroid==2.3.3
astropy==4.0.1.post1
astunparse==1.6.3
atomicwrites==1.3.0
attr==0.3.1
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
bcrypt==3.1.7
beautifulsoup4==4.8.2
billiard==3.3.0.23
bleach==3.1.0
blis==0.4.1
bokeh==2.0.2
boto3==1.12.31
botocore==1.15.31
branca==0.4.0
cachetools==4.0.0
catalogue==1.0.0
celery==3.1.26.post2
certifi==2019.11.28
cffi==1.14.0
cfgv==3.1.0
chainer==7.1.0
chainercv2==0.0.54
chardet==3.0.4
Click==7.0
colorama==0.4.3
configparser==5.0.0
confuse==1.0.0
coverage==5.1
coveralls==2.0.0
cpi==0.1.16
crayons==0.3.0
cv2-tools==2.3.4
cycler==0.10.0
cymem==2.0.3
dash==1.9.0
dash-bootstrap-components==0.8.3
dash-core-components==1.8.0
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.0
datascience==0.15.4
decorator==4.4.1
defusedxml==0.6.0
distlib==0.3.0
Django==3.0.6
django-celery==3.3.1
django-celery-results==1.2.1
django-crispy-forms==1.9.0
djangorestframework==3.11.0
dnspython==1.16.0
docopt==0.6.2
docutils==0.15.2
dominate==2.5.1
easymoney==1.5.0
emoji==0.5.4
entrypoints==0.3
eventlet==0.25.2
filelock==3.0.12
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
Flask-Compress==1.4.0
flask-mongo-session==1.2.0
flask-mongo-sessions==0.2.1
Flask-PyMongo==2.3.0
folium==0.10.1
funcy==1.14
future==0.18.2
Fuzzy==1.2.2
fuzzywuzzy==0.18.0
gast==0.3.3
gensim==3.8.1
geographiclib==1.50
geopy==1.22.0
gevent==20.5.0
google-api-core==1.16.0
google-auth==1.12.0
google-auth-oauthlib==0.4.1
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-cloud-translate==2.0.1
google-pasta==0.2.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
googletrans==2.4.0
graphviz==0.13.2
greenlet==0.4.15
grpcio==1.26.0
gunicorn==20.0.4
h5py==2.10.0
htmlmin==0.1.12
identify==1.4.15
idna==2.8
ImageHash==4.0
imagesize==1.2.0
ipykernel==5.1.4
ipython==7.12.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jmespath==0.9.5
joblib==0.14.1
json5==0.9.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kaggle==1.5.6
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
kmodes==0.10.2
kombu==3.0.37
lazy-object-proxy==1.4.3
llvmlite==0.31.0
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mime==0.1.0
missingno==0.4.2
mistune==0.8.4
mlxtend==0.17.2
monotonic==1.5
more-itertools==8.2.0
murmurhash==1.0.2
nameparser==1.0.6
nbconvert==5.6.1
nbformat==5.0.4
networkx==2.4
nltk==3.4.5
nodeenv==1.3.5
nose==1.3.7
notebook==6.0.3
numba==0.48.0
numexpr==2.7.1
numpy==1.18.1
oauthlib==3.1.0
opencv-python==4.2.0.32
opt-einsum==3.2.0
Orange3-Associate==1.1.5
packaging==20.3
pandas==1.0.3
pandas-datareader==0.8.1
pandas-profiling==2.5.0
pandocfilters==1.4.2
parso==0.6.1
pathtools==0.1.2
patsy==0.5.1
phik==0.9.9
pickleshare==0.7.5
Pillow==7.0.0
pipenv==2018.11.26
pipreqs==0.4.10
plac==1.1.3
plotly==4.5.1
pluggy==0.13.1
pre-commit==2.3.0
preshed==3.0.2
prince==0.7.0
prometheus-client==0.7.1
prompt-toolkit==3.0.3
protobuf==3.11.3
psycopg2==2.8.5
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycairo==1.19.1
pyclustering==0.9.3.1
pycountry==19.8.18
pycparser==2.20
pydot==1.4.1
pydotplus==2.0.2
Pygments==2.5.2
pyjarowinkler==1.8
pyLDAvis==2.1.2
pylint==2.4.4
pymining==0.2
pymongo==3.10.1
pyparsing==2.4.6
pyrsistent==0.15.7
PySocks==1.7.1
pytest==5.4.1
pytest-pylint==0.15.1
python-constraint==1.4.0
python-dateutil==2.8.1
python-igraph==0.7.1.post6
python-Levenshtein==0.12.0
python-slugify==4.0.0
pytz==2019.3
PyWavelets==1.1.1
pywin32==227
pywinpty==0.5.7
pyxDamerauLevenshtein==1.5.3
PyYAML==5.3
pyzmq==18.1.1
qtconsole==4.6.0
rdflib==5.0.0
redis==3.4.1
regex==2020.2.20
requests==2.22.0
requests-oauthlib==1.3.0
retrying==1.3.3
rq==1.3.0
rsa==4.0
s3transfer==0.3.3
scikit-learn==0.23.0
scipy==1.4.1
seaborn==0.10.0
selenium==3.141.0
Send2Trash==1.5.0
silpa-common==0.3
simplekv==0.14.1
six==1.14.0
sklearn==0.0
smart-open==1.10.0
snowballstemmer==2.0.0
soundex==1.1.3
soupsieve==1.9.5
spacy==2.2.4
SPARQLWrapper==1.8.5
Sphinx==3.0.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
SQLAlchemy==1.3.17
sqlparse==0.3.1
srsly==1.0.2
stats==0.1.2a0
statsmodels==0.11.1
style==1.1.0
table==0.0.4
tangled-up-in-unicode==0.0.3
tensorboard==2.2.0
tensorboard-plugin-wit==1.6.0.post2
tensorflow==1.14
tensorflow-estimator==2.2.0rc0
tensorflow-gpu-estimator==2.1.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
text-unidecode==1.3
textblob==0.15.3
thinc==7.4.0
threadpoolctl==2.0.0
titlecase==0.12.0
toml==0.10.0
tornado==6.0.3
tox==3.15.0
tqdm==4.42.0
traitlets==4.3.3
translate==3.5.0
tweepy==3.8.0
typing-extensions==3.7.4.1
update==0.0.1
urllib3==1.24.3
urllib5==5.0.0
vaderSentiment==3.3.1
vine==1.3.0
virtualenv==20.0.20
virtualenv-clone==0.5.3
visions==0.2.2
visitor==0.1.3
waitress==1.4.3
wasabi==0.6.0
watchdog==0.10.2
wbdata==0.2.7
wcwidth==0.1.8
webdriver-manager==2.5.2
webencodings==0.5.1
Werkzeug==0.16.1
whoswho==0.2.1
widgetsnbextension==3.5.1
wordcloud==1.7.0
wrapt==1.11.2
WTForms==2.3.1
xgboost==1.0.2
xlrd==1.2.0
yarg==0.1.9
